// apps/web/src/shared/api/generated/model/deliveryReadControllerInTransitListFilter.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type DeliveryReadControllerInTransitListFilter =
  (typeof DeliveryReadControllerInTransitListFilter)[keyof typeof DeliveryReadControllerInTransitListFilter];

export const DeliveryReadControllerInTransitListFilter = {
  all: "all",
  ordered: "ordered",
  in_transit: "in_transit",
  ready_for_pickup: "ready_for_pickup",
  arriving_soon: "arriving_soon",
  this_week: "this_week",
  delayed: "delayed",
  pickup_expiring: "pickup_expiring",
  awaiting_dispatch: "awaiting_dispatch",
  unassigned: "unassigned",
  no_delivery_date: "no_delivery_date",
  has_tracking_number: "has_tracking_number",
  without_tracking_number: "without_tracking_number",
  has_tracking_url: "has_tracking_url",
  without_tracking_url: "without_tracking_url",
  has_price: "has_price",
  without_price: "without_price",
} as const;